server/requirements.txt
backports.tarfile==1.2.0
channels-redis==4.2.1
django-cors-headers==4.6.0
django-ninja==1.3.0
email-validator==2.2.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
packaging==24.2
pillow==11.0.0
platformdirs==4.2.2
pyopenssl==24.3.0
python-dotenv==1.0.1
service-identity==24.2.0
tomli==2.0.1
zope.interface==7.2
